requirements.txt
ailment==9.2.25
angr==9.2.25
angrop==9.2.7
archinfo==9.2.25
bcrypt==4.0.1
bitstring==4.0.1
cachetools==5.2.0
capstone==4.0.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
claripy==9.2.25
cle==9.2.25
colored-traceback==0.3.0
CppHeaderParser==2.7.4
cryptography==38.0.3
decorator==5.1.1
dpkt==1.9.8
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
intervaltree==3.1.0
itanium-demangler==1.1
Mako==1.2.4
MarkupSafe==2.1.1
mpmath==1.2.1
mulpyplexer==0.9
nampa==0.1.1
networkx==2.8.8
packaging==21.3
paramiko==2.12.0
pathlib2==2.3.7.post1
pefile==2022.5.30
plumbum==1.8.0
ply==3.11
progressbar2==4.2.0
protobuf==4.21.9
psutil==5.9.4
pwntools==4.8.0
pycparser==2.21
pyelftools==0.29
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
PySMT==0.9.6.dev21
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.4.5
pyvex==9.2.25
requests==2.28.1
ROPGadget==7.1
rpyc==5.2.3
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
sympy==1.11.1
tqdm==4.64.1
unicorn==1.0.2rc4
urllib3==1.26.12
z3-solver==4.10.2.0
